+++ CHANGELOG.md
@@ -41,6 +41,15 @@ in a future release.
   - Added new resource SqlServerProtocolTcpIp ([issue #1378](https://github.com/dsccommunity/SqlServerDsc/issues/1378)).
   - Fixing a problem with the latest ModuleBuild 1.7.0 that breaks the CI
     pipeline.
+- SqlServerDsc.Common
+  - Added function `Import-Assembly` that can help import an assembly
+    into the PowerShell session.
+  - Prepared unit tests to support Pester 5 so a minimal conversation
+    is only needed later.
+  - Updated `Import-SQLPSModule` to better support unit tests.
+- CommonTestHelper
+  - Added the functions `Get-InvalidOperationRecord` and `Get-InvalidResultRecord`
+    that is needed for evaluate localized error message strings for unit tests.
 - SqlServerEndpoint
   - BREAKING CHANGE: A new required property `EndpointType` was added to
     support different types of endpoints in the future. For now the only